Text2image/requirements.txt
pillow
pandas
matplotlib
numpy
googletrans==4.0.0-rc1
requests
